#249: Update documentation pages for examples
#249: Add Traversal example to list
