+++ docs/md/checkpoint_learn_example_2.md
@@ -3,5 +3,6 @@
 The full code for this *checkpoint* example can be found here:
 `examples/checkpoint_example_2.cc`
 
+\subsubsection ex Example source code:
 \snippet examples/checkpoint_example_2.cc Serialize custom structure
