+++ docs/md/checkpoint_learn_example_2_nonintrusive.md
@@ -1,7 +1,10 @@
 \page ckpt_learn_ex2_nonintrusive Non-Intrusive Program Example 2
 
+\attention All Non-Intrusive serialize methods <b>MUST</b> be placed in the namespace of type which they serialize.
+
 The full code for this *checkpoint* example can be found here:
 `examples/checkpoint_example_2_nonintrusive.cc`
 
+\subsubsection ex Example source code:
 \snippet examples/checkpoint_example_2_nonintrusive.cc Non-Intrusive Serialize custom structure
