+++ docs/md/checkpoint_learn_example_3.md
@@ -3,4 +3,5 @@
 The full code for this *checkpoint* example can be found here:
 `examples/checkpoint_example_3.cc`
 
+\subsubsection ex3 Example source code:
 \snippet examples/checkpoint_example_3.cc Serialize constraints constructor destructor
