+++ docs/md/checkpoint_learn_example_3_nonintrusive.md
@@ -1,6 +1,9 @@
 \page ckpt_learn_ex3_nonintrusive Non-Intrusive Program Example 3
 
+\attention All Non-Intrusive serialize methods <b>MUST</b> be placed in the namespace of type which they serialize.
+
 The full code for this *checkpoint* example can be found here:
 `examples/checkpoint_example_3_nonintrusive.cc`
 
+\subsubsection ex Example source code:
 \snippet examples/checkpoint_example_3_nonintrusive.cc Non-Intrusive Serialize constraints constructor destructor
