+++ docs/md/checkpoint_learn_example_polymorphic_macro_nonintrusive.md
@@ -1,6 +1,9 @@
 \page ckpt_learn_example_polymorphic_macro_nonintrusive Non-Intrusive Polymorphic Serialization Example w/Macros
 
+\attention All Non-Intrusive serialize methods <b>MUST</b> be placed in the namespace of type which they serialize.
+
 The full code for this *checkpoint* example can be found here:
 `examples/checkpoint_example_polymorphic_macro_nonintrusive.cc`
 
+\subsubsection ex Example source code:
 \snippet examples/checkpoint_example_polymorphic_macro_nonintrusive.cc Non-Intrusive Serialize polymorphic macro
