+++ docs/md/checkpoint_learn_example_to_file.md
@@ -1,8 +1,10 @@
 \page ckpt_learn_ex_tofile Program Example Serialize To File
 
-The full code for this example can be found here:
-`examples/checkpoint_example_to_file.cc`
 One potential application for this approach is to generate a restart file
 for a simulation.
 
+The full code for this example can be found here:
+`examples/checkpoint_example_to_file.cc`
+
+\subsubsection ex Example source code:
 \snippet examples/checkpoint_example_to_file.cc Serialize structure to file
