+++ docs/md/checkpoint_learn_example_traversal.md
@@ -1,0 +1,8 @@
+\page ckpt_learn_ex_traversal Program Example Traversal
+
+The full code for this *checkpoint* example can be found here:
+`examples/checkpoint_example_traversal.cc`
+
+\subsubsection ex Example source code:
+\snippet examples/checkpoint_example_traversal.cc Serialization with custom traverser
+
