+++ docs/md/checkpoint_learn_example_traversal_nonintrusive.md
@@ -1,0 +1,10 @@
+\page ckpt_learn_ex_traversal_nonintrusive Non-Intrusive Program Example Traversal
+
+\attention All Non-Intrusive serialize methods <b>MUST</b> be placed in the namespace of type which they serialize.
+
+The full code for this *checkpoint* example can be found here:
+`examples/checkpoint_example_traversal_nonintrusive.cc`
+
+\subsubsection ex Example source code:
+\snippet examples/checkpoint_example_traversal_nonintrusive.cc Non-Intrusive Serialization with custom traverser
+
