+++ docs/md/checkpoint_learn_examples.md
@@ -8,12 +8,13 @@
 | checkpoint_example_to_file | Serialize / deserialize structure to file | \subpage ckpt_learn_ex_tofile |
 | checkpoint_example_polymorphic | Serialization of polymorphic class hierarchy | \subpage ckpt_learn_example_polymorphic |
 | checkpoint_example_polymorphic_macro | Serialization of polymorphic class hierarchy with macros | \subpage ckpt_learn_example_polymorphic_macro |
+| checkpoint_example_traversal | Serialization with custom traverser | \subpage ckpt_learn_ex_traversal |
 
 &nbsp;
 &nbsp;
 
-\attention All Non-Intrusive serialize methods <b>MUST</b> be placed in the namespace of type which they serialize. 
-Please do not put serialize methods in global magistrate namespace. 
+\attention All Non-Intrusive serialize methods <b>MUST</b> be placed in the namespace of type which they serialize.
+Please do not put serialize methods in global magistrate namespace.
 
 | Non-Intrusive Examples    | Description | Link |
 | ------------------------- | ----------- | ---- |
@@ -23,4 +24,4 @@ Please do not put serialize methods in global magistrate namespace.
 | checkpoint_example_to_file | Serialize / deserialize structure to file | \subpage ckpt_learn_ex_tofile_nonintrusive |
 | checkpoint_example_polymorphic | Serialization of polymorphic class hierarchy | \subpage ckpt_learn_example_polymorphic_nonintrusive |
 | checkpoint_example_polymorphic_macro | Serialization of polymorphic class hierarchy with macros | \subpage ckpt_learn_example_polymorphic_macro_nonintrusive |
-
+| checkpoint_example_traversal | Serialization with custom traverser | \subpage ckpt_learn_ex_traversal_nonintrusive |
