+++ docs/shared/checkpoint_learn_example_1_nonintrusive.md
@@ -1,6 +1,9 @@
 \page ckpt_learn_ex1_nonintrusive Non-Intrusive Program Example 1
 
+\attention All Non-Intrusive serialize methods <b>MUST</b> be placed in the namespace of type which they serialize.
+
 The full code for this *checkpoint* example can be found here:
 `examples/checkpoint_example_1_nonintrusive.cc`
 
+\subsubsection ex1 Example source code:
 \snippet examples/checkpoint_example_1_nonintrusive.cc Non-Intrusive serialize custom structure
