+++ examples/checkpoint_example_1_nonintrusive.cc
@@ -2,7 +2,7 @@
 //@HEADER
 // *****************************************************************************
 //
-//                    checkpoint_example_1_nonnonintrusive.cc
+//                    checkpoint_example_1_nonintrusive.cc
 //                 DARMA/checkpoint => Serialization Library
 //
 // Copyright 2019 National Technology & Engineering Solutions of Sandia, LLC
@@ -48,29 +48,24 @@
 #include <cstdio>
 
 // \brief Namespace containing type which will be serialized
-namespace magistrate::nonintrusive::examples {
+namespace magistrate { namespace nonintrusive { namespace examples {
 
 // \brief Simple structure with three variables of built-in types
 struct MyTest3 {
   int a = 1, b = 2 , c = 3;
 
   // \brief Printing function unto the standard display
-
-  /** @fn void MyTest3::print()
-  *  @brief A member function.
-  *  @param a: input parameter.
-  */
   void print() {
     printf("MyTest3: a=%d, b=%d, c=%d\n", a, b, c);
   }
 };
 
-} // end namespace magistrate::nonintrusive::examples
+}}} // end namespace magistrate::nonintrusive::examples
 
-// \brief Function to serialize the MyTest3 structure. 
+// \brief Function to serialize the MyTest3 structure.
 // In Non-Intrusive way, this function needs to be placed in the namespace
 // of the type which will be serialized.
-namespace magistrate::nonintrusive::examples {
+namespace magistrate { namespace nonintrusive { namespace examples {
   // \brief Templated function for serializing/deserializing
   // a variable of type `MyTest3`. Non-nonintrusive version of the function
   // placed outside of `MyTest3` structure.
@@ -98,7 +93,7 @@ namespace magistrate::nonintrusive::examples {
     s | my_test3.c;
   }
 
-} // end namespace magistrate::nonintrusive::examples
+}}} // end namespace magistrate::nonintrusive::examples
 
 int main(int, char**) {
   using namespace magistrate::nonintrusive::examples;
