+++ examples/checkpoint_example_2_nonintrusive.cc
@@ -2,7 +2,7 @@
 //@HEADER
 // *****************************************************************************
 //
-//                   checkpoint_example_2_nonnonintrusive.cc
+//                   checkpoint_example_2_nonintrusive.cc
 //                 DARMA/checkpoint => Serialization Library
 //
 // Copyright 2019 National Technology & Engineering Solutions of Sandia, LLC
@@ -48,7 +48,7 @@
 #include <cstdio>
 
 // \brief Namespace containing types which will be serialized
-namespace magistrate::nonintrusive::examples {
+namespace magistrate { namespace nonintrusive { namespace examples {
 
 // \struct MyTest2
 // \brief Simple structure with one variable of built-in type
@@ -90,11 +90,12 @@ struct MyTest {
   }
 };
 
-} // end namespace magistrate::nonintrusive::examples
+}}} // end namespace magistrate::nonintrusive::examples
 
 // \brief In Non-Intrusive way, serialize function needs to be placed in the namespace
 // of the type which will be serialized.
-namespace magistrate::nonintrusive::examples {
+namespace magistrate { namespace nonintrusive { namespace examples {
+
 // \brief Templated function for serializing/deserializing
 // a variable of type `MyTest`
 //
@@ -131,7 +132,7 @@ void serialize(Serializer& s, MyTest2& my_test2) {
   s | my_test2.c;
 }
 
-} // end namespace magistrate::nonintrusive::examples
+}}} // end namespace magistrate::nonintrusive::examples
 
 int main(int, char**) {
   using namespace magistrate::nonintrusive::examples;
