+++ examples/checkpoint_example_traversal_nonintrusive.cc
@@ -2,7 +2,7 @@
 //@HEADER
 // *****************************************************************************
 //
-//                           checkpoint_traversal.cc
+//                   checkpoint_traversal_nonintrusive.cc
 //                 DARMA/checkpoint => Serialization Library
 //
 // Copyright 2019 National Technology & Engineering Solutions of Sandia, LLC
@@ -41,12 +41,15 @@
 //@HEADER
 */
 
+/// [Non-Intrusive Serialization with custom traverser]
+
 #include <checkpoint/checkpoint.h>
 
 #include <cstdio>
 #include <string>
 
-namespace magistrate::intrusive::examples {
+// \brief Namespace containing type which will be serialized
+namespace magistrate { namespace nonintrusive { namespace examples {
 
 struct TestObject {
 
@@ -63,9 +66,7 @@ struct TestObject {
   }
 
   template <typename Serializer>
-  void serialize(Serializer& s) {
-    s | a | b | vec1 | vec2 | vec3;
-  }
+  friend void serialize(Serializer& s, TestObject& obj);
 
 private:
   int a = 29;
@@ -75,7 +76,20 @@ private:
   std::vector<std::string> vec3;
 };
 
-} // end namespace magistrate::intrusive::examples
+}}} // end namespace magistrate::nonintrusive::examples
+
+// \brief In Non-Intrusive way, serialize functionality needs to be placed in the namespace
+// of the type which will be serialized.
+namespace magistrate { namespace nonintrusive { namespace examples {
+
+template <typename Serializer>
+void serialize(Serializer& s, TestObject& obj) {
+  s | obj.a;
+  s | obj.b;
+  s | obj.vec1;
+  s | obj.vec2;
+  s | obj.vec3;
+}
 
 /// Custom traverser for printing raw bytes
 struct PrintBytesTraverse : checkpoint::Serializer {
@@ -135,8 +149,10 @@ struct TypedTraverse : checkpoint::Serializer {
   }
 };
 
+}}} // end namespace magistrate::nonintrusive::examples
+
 int main(int, char**) {
-  using namespace magistrate::intrusive::examples;
+  using namespace magistrate::nonintrusive::examples;
 
   TestObject my_obj(TestObject::MakeTag{});
 
@@ -149,3 +165,5 @@ int main(int, char**) {
 
   return 0;
 }
+
+/// [Non-Intrusive Serialization with custom traverser]
